refactor: move Approximator to core module
